requirements.txt
Flask==2.3.2
Flask-Cors==4.0.0
psycopg2-binary==2.9.7
pdfplumber==0.10.2
python-dotenv==1.0.0
Werkzeug==2.3.6
PyJWT==2.8.0
pandas==2.0.3
streamlit==1.25.0
openpyxl==3.1.2